capping

+++ retailer_to_sp/api/v1/views.py
@@ -439,15 +439,20 @@ class AddToCart(APIView):
                     cart.save()
 
                 if capping.capping_qty > ordered_qty:
-                    if int(qty) == 0:
-                        if CartProductMapping.objects.filter(cart=cart, cart_product=product).exists():
-                            CartProductMapping.objects.filter(cart=cart, cart_product=product).delete()
+                    if capping.capping_qty > int(qty):
+                        if int(qty) == 0:
+                            if CartProductMapping.objects.filter(cart=cart, cart_product=product).exists():
+                                CartProductMapping.objects.filter(cart=cart, cart_product=product).delete()
 
+                        else:
+                            cart_mapping, _ = CartProductMapping.objects.get_or_create(cart=cart, cart_product=product)
+                            cart_mapping.qty = qty
+                            cart_mapping.no_of_pieces = int(qty) * int(product.product_inner_case_size)
+                            cart_mapping.capping_error_msg = ''
+                            cart_mapping.save()
                     else:
                         cart_mapping, _ = CartProductMapping.objects.get_or_create(cart=cart, cart_product=product)
-                        cart_mapping.qty = qty
-                        cart_mapping.no_of_pieces = int(qty) * int(product.product_inner_case_size)
-                        cart_mapping.capping_error_msg = ''
+                        cart_mapping.capping_error_msg = 'Raj Shekhar Singh'
                         cart_mapping.save()
                 else:
                     cart_mapping, _ = CartProductMapping.objects.get_or_create(cart=cart, cart_product=product)
